Hash Function 1
Hash Function 1

Key:
To get the key this program is adding up all of the ascii values of each letter, and then modding by the size of the table.

Collisions:
Collisions are delt with by adding the data to a sublist at the location of the key within the table.

Comments:
there are a lot of collisions and wasted space, as well as it looks like none of the keys are populating the second half of the table, so I am going to try to find another way to generate the keys that will make larger numbers.