Hash Function 3
Hash Function 3

Key:
To create the key this time I tried to do something similar to last time, but I added onto it. Now after the previous hash, it adds the ascii value of each letter to the third power to the number, and then floor divides by another large prime number.

Collisions:
Collisions are still delt with in the same way, because I want to be able to find the best way to generate a key first, then try to find a more efficient way to deal with collisions using that key function.

Comments:
This function was actually less effective then the previous one which shocked me. I had expected that doing more manipulation to the keys with prime numbers would decrease collisions, but it seems that at a certain point the more you add it starts going in the other direction and gets worse.