Hash Function 4
Hash Function 4

Key:
This time I took inspiration from a commonly used hash function that adds the ascii value times 31 to the power of the length of the string minus the index of the character minus 1

Collisions:
Still using lists for collisions.

Comments:
THis one was slightly more space efficient, biut took drastically longer then the other versions. I am going to make a new way to deal with collisions and test it with both this key generation function, and the function from the second commit version to see how they compare.