Hash Function 5
Hash Function 5

Key:
I used the same key generator as in hash function 4.

Collisions:
To deal with collisions, I made the array larger by a factor of 1.3, and made it so that each time a value collided, it moved to the next spot and checked if it could fit there, and so on.

Comments:
This program was slower then the previous version where it used an array of all of the data that collided, but it took less space in memory, so it might be better depending on what is more important, time or space.