Hash Function 6
Hash Function 6

Key:
I used the same key generator as in hash function 2

Collisions:
This one deals with collisions the same as hash function 5

Comments:
This program was faster then hash program 5, and uses the same amount of memory, but it has a lot more collisions. This means that in a use case where it is expensive to read the data, it would be less effective, and hash function 5 would be better, but in a case where reading data is cheaper then this one would be faster.